consolidate storage pg pools and dedupe fs usage helpers
